# app/docs/swagger.yaml
definitions:
...
  response.ListData:
    properties:
      rows: {}
      total:
        type: integer
    type: object
  response.ResCode:
    enum:
    - 200
    - 401
    - 403
    - 412
    - 500
    - 600
    type: integer
    x-enum-comments:
      Error: 系统异常
      Forbidden: 没有权限
      Parameter: 参数错误
      Success: 成功
      Unauthorized: token失效
      Waring: 详情看msg
    x-enum-varnames:
    - Success
    - Unauthorized
    - Forbidden
    - Parameter
    - Error
    - Waring
  response.ResponseData:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/response.ResCode'
        description: 相应状态码
      data:
        description: 数据
      msg:
        description: 提示信息
        type: string
    type: object
...
  systemModels.SysRoleDML:
    properties:
      RoleId:
        example: "0"
        type: string
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      menuIds:
        items:
          type: string
        type: array
      remark:
        type: string
      roleKey:
        type: string
      roleName:
        type: string
      roleSort:
        type: integer
      status:
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
    type: object
  systemModels.SysRoleVo:
    properties:
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      delFlag:
        type: string
      remark:
        type: string
      roleId:
        example: "0"
        type: string
      roleKey:
        type: string
      roleName:
        type: string
      roleSort:
        type: integer
      status:
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
    type: object
...
  systemModels.SysUserRole:
    properties:
      roleId:
        example: "0"
        type: string
      userId:
        example: "0"
        type: string
    type: object
  systemModels.SysUserVo:
    properties:
      avatar:
        type: string
      createBy:
        description: 创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      dataScope:
        type: string
      delFlag:
        type: string
      deptId:
        example: "0"
        type: string
      deptName:
        type: string
      email:
        type: string
      leader:
        type: string
      nickName:
        type: string
      phonenumber:
        type: string
      remark:
        type: string
      roleId:
        type: integer
      sex:
        type: string
      status:
        type: string
      updateBy:
        description: 修改人
        type: integer
      updateTime:
        description: 修改时间
        type: string
      userId:
        example: "0"
        type: string
      userName:
        type: string
    type: object
...
info:
  contact:
    email: zhao_402295440@126.com
    name: danny
  description: baize接口文档
  title: baize
  version: 2.0.x
paths:
...
  /system/role:
    post:
      description: 添加角色
      parameters:
      - description: 公司信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysRoleDML'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 添加角色
      tags:
      - 角色相关
    put:
      description: 修改角色
      parameters:
      - description: 公司信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysRoleDML'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 修改角色
      tags:
      - 角色相关
  /system/role/{roleId}:
    get:
      description: 根据角色ID获取角色信息
      parameters:
      - description: roleId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/systemModels.SysRoleVo'
              type: object
      security:
      - BearerAuth: []
      summary: 根据角色ID获取角色信息
      tags:
      - 角色相关
  /system/role/{rolesIds}:
    delete:
      description: 删除角色
      parameters:
      - collectionFormat: csv
        description: rolesIds
        in: path
        items:
          type: string
        name: rolesIds
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - 角色相关
  /system/role/authUser/cancel:
    put:
      description: 取消用户角色
      parameters:
      - description: 用户id角色id
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysUserRole'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 取消用户角色
      tags:
      - 角色相关
  /system/role/authUser/cancelAll:
    put:
      description: 取消角色授权用户
      parameters:
      - description: 角色Id
        in: query
        name: roleId
        required: true
        type: string
      - collectionFormat: csv
        description: 用户Ids
        in: query
        items:
          type: string
        name: userIds
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 取消角色授权用户
      tags:
      - 角色相关
  /system/role/authUser/selectAll:
    put:
      description: 角色授权用户
      parameters:
      - description: 角色Id
        in: query
        name: roleId
        required: true
        type: string
      - collectionFormat: csv
        description: 用户Ids
        in: query
        items:
          type: string
        name: userIds
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 角色授权用户
      tags:
      - 角色相关
  /system/role/authUser/unallocatedList:
    put:
      description: 查询角色未授权用户列表查询
      parameters:
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: phonenumber
        type: string
      - in: query
        name: roleId
        required: true
        type: string
      - in: query
        name: userName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListData'
                  - properties:
                      Rows:
                        items:
                          $ref: '#/definitions/systemModels.SysUserVo'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询角色未授权用户列表查询
      tags:
      - 角色相关
  /system/role/changeStatus:
    get:
      description: 查询角色授权用户列表查询
      parameters:
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: phonenumber
        type: string
      - in: query
        name: roleId
        required: true
        type: string
      - in: query
        name: userName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListData'
                  - properties:
                      Rows:
                        items:
                          $ref: '#/definitions/systemModels.SysUserVo'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询角色授权用户列表查询
      tags:
      - 角色相关
    put:
      description: 修改角色状态
      parameters:
      - description: 公司信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/systemModels.SysRoleDML'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.ResponseData'
      security:
      - BearerAuth: []
      summary: 修改角色状态
      tags:
      - 角色相关
  /system/role/list:
    get:
      description: 查询角色列表查询
      parameters:
      - in: query
        name: beginTime
        type: string
      - in: query
        name: endTime
        type: string
      - description: 排序规则  降序desc   asc升序
        in: query
        name: isAsc
        type: string
      - description: 排序字段
        in: query
        name: orderBy
        type: string
      - default: 1
        description: 第几页
        in: query
        name: pageNum
        type: integer
      - default: 10000
        description: 数量
        in: query
        name: pageSize
        type: integer
      - in: query
        name: roleKey
        type: string
      - in: query
        name: roleName
        type: string
      - in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseData'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListData'
                  - properties:
                      Rows:
                        items:
                          $ref: '#/definitions/systemModels.SysRoleVo'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询角色列表查询
      tags:
      - 角色相关
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"